Updater: Fix pending reboot state not shown
When the 'Auto delete updates' preference is enabled or with
local updates, the updater removes the update record from the
database immediately after a successful installation.
This leads to pending reboot state not shown.

To resolve this, this change defers the automatic deletion of
the update record until after the device has rebooted successfully.

Change-Id: I3f4037cbd8f4fd5145d14ab7301b0e99913ddbe0